DEVEL: Changing Oauth2 ctor to one or two overloads, adding idiotproofs
json2sqlite is having idiotproofing added.

+++ json2sqlite/json2sqlite.cpp
@@ -7,55 +7,63 @@ Json2Sqlite::Json2Sqlite()
 
 QJsonObject Json2Sqlite::loadSettings(QString dbPath, QJsonObject jsonSettings)
 {
-    if(!doesDatabaseExist(dbPath))
-        makeInitalSettingsDatabase(dbPath);
-
-    QSqlDatabase database;
-    database = database.addDatabase("QSQLITE", "settings");
-    database.setDatabaseName(dbPath);
-
-    if(!database.open())
+    if(!dbPath.isEmpty() && !jsonSettings.isEmpty())
     {
-        qDebug() << "db failed to load in OAuth2 loadSettings"
-                 << "All settings values are null!";
+        if(!doesDatabaseExist(dbPath))
+            makeInitalSettingsDatabase(dbPath);
+
+        QSqlDatabase database;
+        database = database.addDatabase("QSQLITE", "settings");
+        database.setDatabaseName(dbPath);
+
+        if(!database.open())
+        {
+            qDebug() << "db failed to load in OAuth2 loadSettings"
+                     << "All settings values are null!";
+
+            return jsonSettings;
+        }
+
+        QSqlQuery query(database);
+        QString queryString  = "SELECT * FROM settings WHERE key IN (";
+
+        for(auto key: jsonSettings.keys()){
+            queryString.append("'" + key + "', ");
+        }
+        queryString.remove((queryString.length() - 2), 2);
+        queryString.append(")");
+
+        query.prepare(queryString);
+
+        if(query.exec())
+        {
+            qDebug() << "Json2Sqlite::loadSettings query success.";
+
+            while(query.next())
+            {
+                if(query.value("isJsonArray").toBool())
+                    jsonSettings[query.value("key").toString()] = QJsonDocument::fromJson(query.value("value").toString().toUtf8()).array();
+                else
+                    jsonSettings[query.value("key").toString()] = query.value("value").toJsonValue();
+            }
+        }
+        else
+        {
+            qDebug() << "Json2Sqlite::loadSettings ERROR: " << query.lastError();
+        }
+        query.clear();
+
+        database.close();
+        database = QSqlDatabase();
+        database.removeDatabase("settings");
 
         return jsonSettings;
     }
-
-    QSqlQuery query(database);
-    QString queryString  = "SELECT * FROM settings WHERE key IN (";
-
-    for(auto key: jsonSettings.keys()){
-        queryString.append("'" + key + "', ");
-    }
-    queryString.remove((queryString.length() - 2), 2);
-    queryString.append(")");
-
-    query.prepare(queryString);
-
-    if(query.exec())
-    {
-        qDebug() << "OAuth2::loadSettings query success.";
-
-        while(query.next())
-        {
-            if(query.value("isJsonArray").toBool())
-                jsonSettings[query.value("key").toString()] = QJsonDocument::fromJson(query.value("value").toString().toUtf8()).array();
-            else
-                jsonSettings[query.value("key").toString()] = query.value("value").toJsonValue();
-        }
-    }
     else
     {
-        qDebug() << "OAuth2::loadSettings ERROR: " << query.lastError();
+        qDebug() << "Json2Sqlite::loadSettings error, dbPath or JsonObject is null";
+        return QJsonObject();
     }
-    query.clear();
-
-    database.close();
-    database = QSqlDatabase();
-    database.removeDatabase("settings");
-
-    return jsonSettings;
 }
 
 QJsonArray Json2Sqlite::loadArray(QString dbPath, QString table)
@@ -67,58 +75,65 @@ QJsonArray Json2Sqlite::loadArray(QString dbPath, QString table)
 
 bool Json2Sqlite::saveSettings(QString dbPath, QJsonObject jsonSettings)
 {
-    if(!doesDatabaseExist(dbPath))
-        makeInitalSettingsDatabase(dbPath);
-
     bool success = false;
-    QString queryString = "INSERT or REPLACE into settings VALUES ";
 
-    QSqlDatabase database;
+    if(!dbPath.isEmpty()){
+        if(!doesDatabaseExist(dbPath))
+            makeInitalSettingsDatabase(dbPath);
 
-    database = database.addDatabase("QSQLITE", "settings");
-    database.setDatabaseName(dbPath);
+        QString queryString = "INSERT or REPLACE into settings VALUES ";
 
-    if(!database.open())
-    {
-        qDebug() << "db failed to load in OAuth2 loadSettings"
-                 << "All settings values are null!";
+        QSqlDatabase database;
 
-        return false;
-    }
+        database = database.addDatabase("QSQLITE", "settings");
+        database.setDatabaseName(dbPath);
 
-    QSqlQuery query(database);
-
-    for(auto key: jsonSettings.keys())
-    {
-        if(jsonSettings[key].isArray())
+        if(!database.open())
         {
-            QJsonDocument arrayToString;
-            arrayToString.setArray(jsonSettings[key].toArray());
-            queryString.append("('" + key + "', '" + QString(arrayToString.toJson()) + "', '" + QString::number(true) + "'),");
-        }
-        else{
-            queryString.append("('" + key + "', '" + jsonSettings[key].toString() + "', '" + QString::number(false) + "'),");
-        }
-    }
-    queryString.remove(queryString.length() - 1, 1);
+            qDebug() << "db failed to load in OAuth2 loadSettings"
+                     << "All settings values are null!";
 
-    query.prepare(queryString);
-    if(query.exec())
-    {
-        qDebug() << "OAuth2::saveSettings query success.";
-        success = true;
+            return false;
+        }
+
+        QSqlQuery query(database);
+
+        for(auto key: jsonSettings.keys())
+        {
+            if(jsonSettings[key].isArray())
+            {
+                QJsonDocument arrayToString;
+                arrayToString.setArray(jsonSettings[key].toArray());
+                queryString.append("('" + key + "', '" + QString(arrayToString.toJson()) + "', '" + QString::number(true) + "'),");
+            }
+            else{
+                queryString.append("('" + key + "', '" + jsonSettings[key].toString() + "', '" + QString::number(false) + "'),");
+            }
+        }
+        queryString.remove(queryString.length() - 1, 1);
+
+        query.prepare(queryString);
+        if(query.exec())
+        {
+            qDebug() << "OAuth2::saveSettings query success.";
+            success = true;
+        }
+        else
+        {
+            qDebug() << "OAuth2::saveSettings ERROR: " << query.lastError();
+            success = false;
+        }
+
+        query.clear();
+        database.close();
+        database = QSqlDatabase();
+        database.removeDatabase("settings");
+        return success;
     }
     else
     {
-        qDebug() << "OAuth2::saveSettings ERROR: " << query.lastError();
-        success = false;
+        return success;
     }
-
-    query.clear();
-    database.close();
-    database = QSqlDatabase();
-    database.removeDatabase("settings");
-    return success;
 }
 
 bool Json2Sqlite::saveArray(QString dbPath, QString table, QString primaryKey, QJsonArray jsonArray)
@@ -205,12 +220,10 @@ bool Json2Sqlite::saveArray(QString dbPath, QString table, QString primaryKey, Q
         return false;
     }
 
-
 //    //PREP TABLE FOR NEW ARRAY
 //    //An append function would just add to table...
 //    query.prepare("DELETE FROM " + table);
 //    query.exec();
-
 
     query.prepare(queryString);
 
@@ -230,7 +243,6 @@ bool Json2Sqlite::saveArray(QString dbPath, QString table, QString primaryKey, Q
     database = QSqlDatabase();
     database.removeDatabase("array");
     return success;
-
 }
 
 bool Json2Sqlite::doesDatabaseExist(QString dbPath)
